+++ app/src/main/res/layout/fragment_favlist.xml
@@ -24,7 +24,5 @@
         android:id="@+id/recyclerView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@id/favText" />
 </androidx.constraintlayout.widget.ConstraintLayout>
